Reorganiza estrutura de pastas do contexto de loterias
