+++ src/components/Sidebar.tsx
@@ -1,14 +1,17 @@
-import { Select } from './Select';
-import logoImg from '../assets/images/logo.svg';
-import '../styles/sidebar.scss';
-import { useContext } from 'react';
-import { LotteriesContext } from '../contexts/LotteriesContext';
+import { Select } from "./Select";
+import logoImg from "../assets/images/logo.svg";
+import "../styles/sidebar.scss";
+import { useContext } from "react";
+import { LotteriesContext } from "../contexts/LotteriesContext";
 
 export function Sidebar() {
   const { currentLotteryDraw } = useContext(LotteriesContext);
 
   return (
-    <div className="sidebar" style={{ backgroundColor: currentLotteryDraw.color }}>
+    <div
+      className="sidebar"
+      style={{ backgroundColor: currentLotteryDraw.color }}
+    >
       <Select label={currentLotteryDraw.lotteryName} />
 
       <div className="logo">
@@ -18,7 +21,9 @@ export function Sidebar() {
 
       <div className="contest">
         <span>Concurso</span>
-        <span>{currentLotteryDraw.contestId} - {currentLotteryDraw.date}</span>
+        <span>
+          {currentLotteryDraw.contestId} - {currentLotteryDraw.date}
+        </span>
       </div>
     </div>
   );
