+++ src/contexts/LotteriesContext/types.d.ts
@@ -1,0 +1,13 @@
+interface LotteryOption {
+  name: string;
+  id: number;
+}
+
+interface LotteryDraw {
+  lotteryName: string;
+  lotteryId: number;
+  contestId: string;
+  numbers: Array<string>;
+  date: string;
+  color: string;
+}
